fabric: executor: Refactor into sub-module
